Core: Replace Materia.fromType with MaterialLoader.createMaterialFromType
